+++ libraries/INA236/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.5] - 2025-11-17
+- sync INA226, 0.6.5
+- improvements on ALERT register functions
+- removed FC00 mask in **bool setAlertRegister()**
+- add **uint16_t getAlertRegister()** to get all register bits.
+  - obsoletes getAlertFlag().
+- add **void setAlertLatchEnable(bool latch = false)**
+- add **void setAlertPolarity(bool inverted = false)**
+- add **void getAlertLatchEnable()**
+- add **void getAlertPolarity()**
+- update readme.md
+- minor edits
+
 ## [0.1.4] - 2025-08-18
 - update GitHub actions
 - allow overrule INA236_MINIMAL_SHUNT on command line.